# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CMS Đoàn Hội API",
        "description": "API documentation for CMS Đoàn Hội - Student Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng ký sinh viên mới",
                "description": "Tạo tài khoản sinh viên mới với thông tin đầy đủ",
                "operationId": "08136088c4862c313b35d6518bab8d3a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "student_id",
                                    "class",
                                    "faculty",
                                    "course",
                                    "date_of_birth",
                                    "gender"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Nguyễn Văn A"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "student@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "student_id": {
                                        "type": "string",
                                        "example": "SV001"
                                    },
                                    "class": {
                                        "type": "string",
                                        "example": "CNTT01"
                                    },
                                    "faculty": {
                                        "type": "string",
                                        "example": "Công nghệ thông tin"
                                    },
                                    "course": {
                                        "type": "string",
                                        "example": "2024"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2000-01-01"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female"
                                        ],
                                        "example": "male"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Đăng ký thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đăng ký thành công"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "full_name": {
                                                            "type": "string",
                                                            "example": "Nguyễn Văn A"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "student@example.com"
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "example": "student"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "student": {
                                                    "properties": {
                                                        "student_id": {
                                                            "type": "string",
                                                            "example": "SV001"
                                                        },
                                                        "class": {
                                                            "type": "string",
                                                            "example": "CNTT01"
                                                        },
                                                        "faculty": {
                                                            "type": "string",
                                                            "example": "Công nghệ thông tin"
                                                        },
                                                        "course": {
                                                            "type": "string",
                                                            "example": "2024"
                                                        },
                                                        "date_of_birth": {
                                                            "type": "string",
                                                            "example": "2000-01-01T00:00:00.000000Z"
                                                        },
                                                        "gender": {
                                                            "type": "string",
                                                            "example": "male"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "0123456789"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|abc123..."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng nhập sinh viên",
                "description": "Đăng nhập với email và mật khẩu để nhận token",
                "operationId": "99d66635c4992aeaa6aa44ff653d0563",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "student@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đăng nhập thành công"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "full_name": {
                                                            "type": "string",
                                                            "example": "Nguyễn Văn A"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "student@example.com"
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "example": "student"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "student": {
                                                    "properties": {
                                                        "student_id": {
                                                            "type": "string",
                                                            "example": "SV001"
                                                        },
                                                        "class": {
                                                            "type": "string",
                                                            "example": "CNTT01"
                                                        },
                                                        "faculty": {
                                                            "type": "string",
                                                            "example": "Công nghệ thông tin"
                                                        },
                                                        "course": {
                                                            "type": "string",
                                                            "example": "2024"
                                                        },
                                                        "date_of_birth": {
                                                            "type": "string",
                                                            "example": "2000-01-01T00:00:00.000000Z"
                                                        },
                                                        "gender": {
                                                            "type": "string",
                                                            "example": "male"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "0123456789"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "2|abc123..."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Email hoặc mật khẩu không đúng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Email hoặc mật khẩu không đúng"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng xuất sinh viên",
                "description": "Đăng xuất và thu hồi token hiện tại",
                "operationId": "6abf3b64a4bc7838d56346f05a5153af",
                "responses": {
                    "200": {
                        "description": "Đăng xuất thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đăng xuất thành công"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/auth/change-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đổi mật khẩu",
                "description": "Đổi mật khẩu của sinh viên",
                "operationId": "392f5b8efb38f64f049872d18d481cf1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "new_password",
                                    "new_password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "oldpassword123"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    },
                                    "new_password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đổi mật khẩu thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đổi mật khẩu thành công. Vui lòng đăng nhập lại."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Mật khẩu hiện tại không đúng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mật khẩu hiện tại không đúng"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/student/events": {
            "get": {
                "tags": [
                    "Student Events"
                ],
                "summary": "Lấy danh sách sự kiện",
                "description": "Lấy danh sách tất cả sự kiện có thể đăng ký",
                "operationId": "7a8b93212cab50f76e798f1f10f32402",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Lọc theo trạng thái sự kiện",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "upcoming",
                                "ongoing",
                                "completed"
                            ]
                        }
                    },
                    {
                        "name": "union_id",
                        "in": "query",
                        "description": "Lọc theo đoàn hội",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Số trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Số item per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách sự kiện",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "events": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "pagination": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/student/statistics": {
            "get": {
                "tags": [
                    "Student Events"
                ],
                "summary": "Lấy thống kê sinh viên",
                "description": "Lấy thống kê tổng quan của sinh viên",
                "operationId": "f92279dfee94f6739d5bb039f4710d98",
                "responses": {
                    "200": {
                        "description": "Thống kê sinh viên",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}